refactor(search): replace CmdKPalette dialog with ⌘K focus hotkey
Desktop search is now exclusively the always-visible ProjectSearchBar
on project routes. The hotkey changes:

- ⌘/ + Ctrl+/ → ⌘K + Ctrl+K (modern convention; Linear, GitHub,
  Notion, Cursor all use ⌘K)
- Behavior: focus the existing inline input instead of opening a
  separate dialog
- Scope: desktop only — useEffect guards on breakpoint === 'desktop'.
  Mobile + tablet retain the icon-button + MobileSearchSheet flow
  via SearchEntry (unchanged)

Why this shape:
- The previous CmdKPalette dialog duplicated all of ProjectSearchBar's
  UI (popover, results, recents, partial-permission banner). Routing
  the hotkey at the inline input removes ~135 lines of UI duplication
  while preserving every feature.
- Browser conflict: Chrome + Firefox use Ctrl+K (Cmd+K on Mac) for
  the URL bar's search-bar focus. preventDefault overrides it.
- The listener skips if the event target IS the input — pressing
  Cmd+K while typing in our input shouldn't double-fire.

Removed:
- app/features/search/surfaces/CmdKPalette.tsx (entire file)
- 'cmd-k' variant from SearchSurface telemetry union (test fixtures
  rebased to 'project-bar')

Tradeoff: the hotkey only works on /project/:id/* routes on desktop
(where ProjectSearchBar is mounted). Off-project pages have no
hotkey — acceptable, matches the v3 'search is project-scoped'
model.
